CRUD over using rest

+++ mvc/src/main/java/com/web/mvc/service/ProductService.java
@@ -1,5 +1,6 @@
 package com.web.mvc.service;
 
+import com.web.mvc.exception.ProductNotFoundException;
 import com.web.mvc.model.Product;
 import com.web.mvc.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,8 +19,13 @@ public class ProductService {
         return (List<Product>) productRepository.findAll();
     }
 
-    public Optional<Product> getproductById(long id) {
-        return productRepository.findById(id);
+    public Product getproductById(long id) {
+        try {
+            return productRepository.findById(id)
+                    .orElseThrow(() -> new ProductNotFoundException("Product with ID " + id + " not found"));
+        } catch (ProductNotFoundException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     public Product createProduct(Product product) {
@@ -36,7 +42,26 @@ public class ProductService {
                 .orElseThrow(() -> new RuntimeException("Product not found"));
     }
 
-    public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+    public boolean deleteProduct(Long id) {
+        if (productRepository.existsById(id)) {
+            productRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    public List<Product> getProductByName(String name) {
+        List<Product> products = productRepository.findProductByNameIgnoreCase(name);
+
+        if(products.isEmpty()) {
+            try {
+                throw new ProductNotFoundException("Product with name '" + name + "' not found");
+            } catch (ProductNotFoundException e) {
+                throw new RuntimeException(e);
+            }
+        } else {
+            return products;
+        }
     }
 }
